+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - RTT, bytes per second and packet loss information for `RepliconClient` and `ConnectedClients`.
 - `ClientSet::Diagnostics` for systems that collect client diagnostics.
 
+### Fixed
+
+- Sending removals and despawns for hidden entities.
+
 ### Changed
 
 - Make `core::replication::replication_rules::ReplicationRules` public.